# AK: Discard bits from LittleEndianInputBitStream as they are read