Rather than tracking our position in the bit buffer, we can simply shift away the bits that we read. This is mostly for simplicity, but also does help performance a bit.

Using the "enwik8" file as a test (100MB uncompressed, commonly used in benchmarks: https://www.mattmahoney.net/dc/enwik8.zip), compression time decreases from:

    3.96s to 3.79s on Serenity (cold)
    1.08s to 1.04s on Serenity (warm)
    0.83s to 0.82s on Linux

Credit to zaggy for the idea!